Move DrawShaded to kas::draw
