+++ kas-wgpu/src/draw/shaded_round.rs
@@ -8,8 +8,9 @@
 use std::f32::consts::FRAC_PI_2;
 use std::mem::size_of;
 
-use crate::draw::{Colour, Rgb, Vec2};
+use crate::draw::{Rgb, Vec2};
 use crate::shared::SharedState;
+use kas::draw::Colour;
 use kas::geom::{Rect, Size};
 
 /// Offset relative to the size of a pixel used by the fragment shader to
